Clean up auth
