--- lib/tvdbapiv2/sessions/__init__.py
@@ -1,10 +1,0 @@
-# coding=utf-8
-
-"""
-This package provides custom sessions and related utilities.
-
-The sessions package provides custom sessions to extend a requests.Session
-and to persist settings across requests (cookies, auth, proxies).
-
-The sessions.auth package contains custom authentication handlers for Requests.
-"""
