--- lib/tvdbapiv2/sessions/auth/utils/__init__.py
@@ -1,5 +1,0 @@
-# coding=utf-8
-
-"""
-This package provides utilities for use with session authentication.
-"""
