--- lib/tvdbapiv2/tvdb/__init__.py
@@ -1,5 +1,0 @@
-# coding=utf-8
-
-"""
-This package provides an interface to TheTVDB API.
-"""
