--- lib/tvdbapiv2/tvdb/sessions/__init__.py
@@ -1,5 +1,0 @@
-# coding=utf-8
-
-"""
-This package provides a custom session for TheTVDB.
-"""
