--- lib/tvdbapiv2/tvdb/sessions/auth/__init__.py
@@ -1,5 +1,0 @@
-# coding=utf-8
-
-"""
-This package provides a session authentication for TheTVDB.
-"""
